+++ packages/ui/DrawingCanvas/Drawing.module.less
@@ -1,0 +1,6 @@
+.drawing {
+	margin: 0 auto;
+	width: 70vw;
+	height: 70vh;
+	border: 3px solid pink;
+}
